Code from Tiv: stag ability fixes
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@104 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/world/WorldMgr.cs
@@ -460,7 +460,7 @@ namespace DOL.GS
 						if (player.ObjectState != eObjectState.Active)
 							continue;
 
-						if (Environment.TickCount - player.LastNPCUpdate > PLAYER_UPDATE_TIME)
+						if ((uint)Environment.TickCount - player.LastNPCUpdate > PLAYER_UPDATE_TIME)
 						{
 							BitArray carray = player.CurrentUpdateArray;
 							BitArray narray = player.NewUpdateArray;
@@ -491,7 +491,7 @@ namespace DOL.GS
 							}
 							
 							player.SwitchUpdateArrays();
-							player.LastNPCUpdate = Environment.TickCount;
+							player.LastNPCUpdate = (uint)Environment.TickCount;
 						}
 					}
 					int took = Environment.TickCount - start;
